Update documentation2 (#52)
* added new docs version

* changes from documentatio2

+++ src/pages/docs/versions/master/update_usagedocs.sh
@@ -12,9 +12,20 @@ do
     fi
 
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/roadmap.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/roadmap.md 
-    wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/database.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
+    if [[ `wget -S --spider https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/database.md  2>&1 | grep 'HTTP/1.1 200 OK'` ]]; then 
+        wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/database.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
+    fi
+    if [[ `wget -S --spider https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/document_store.md  2>&1 | grep 'HTTP/1.1 200 OK'` ]]; then 
+        wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/document_store.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/document_store.md 
+    fi
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/domain_adaptation.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/domain_adaptation.md 
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/intro.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/intro.md 
+    if [[ `wget -S --spider https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/languages.md  2>&1 | grep 'HTTP/1.1 200 OK'` ]]; then 
+        wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/languages.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/languages.md 
+    fi
+    if [[ `wget -S --spider https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/preprocessing.md  2>&1 | grep 'HTTP/1.1 200 OK'` ]]; then 
+        wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/preprocessing.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/preprocessing.md 
+    fi
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/reader.md -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/reader.md 
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/retriever.md  -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/retriever.md 
     wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/usage/usage/terms.md  -O ./src/pages/docs/versions/master/${i}site/en/usage/usage/terms.md 
@@ -26,12 +37,26 @@ do
 
     sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/roadmap.md
     sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/roadmap.md 
-    sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
-    sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
+    if [[ -f "./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md" ]]; then
+        sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
+        sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/database.md 
+    fi
+    if [[ -f "./src/pages/docs/versions/master/${i}site/en/usage/usage/document_store.md" ]]; then
+        sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/document_store.md 
+        sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/document_store.md 
+    fi
     sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/domain_adaptation.md 
     sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/domain_adaptation.md 
     sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/intro.md 
     sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/intro.md 
+    if [[ -f "./src/pages/docs/versions/master/${i}site/en/usage/usage/languages.md" ]]; then
+        sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/languages.md 
+        sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/languages.md 
+    fi
+    if [[ -f "./src/pages/docs/versions/master/${i}site/en/usage/usage/preprocessing.md" ]]; then
+        sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/preprocessing.md 
+        sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/preprocessing.md 
+    fi
     sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/reader.md 
     sed -i 's/--->/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/reader.md 
     sed -i 's/<!---/---/' ./src/pages/docs/versions/master/${i}site/en/usage/usage/retriever.md
